requirements.txt
pandas==2.2.3
plotly-express==0.4.1
streamlit==1.39.0
